fix errors and add models author and bookseries

+++ src/main/java/com/shadoww/BookLibraryApp/dto/request/BookMarkRequest.java
@@ -9,8 +9,8 @@ import lombok.Setter;
 @Getter
 public class BookMarkRequest {
 
-    private int bookId;
-    private int chapterId;
+    private long bookId;
+    private long chapterId;
     private int paragraph;
